+++ tests/testthat/test-coef_omit.R
@@ -9,7 +9,7 @@ test_that("perl=TRUE allows lookbehind", {
                       coef_omit="^(?!mpg).*vs",
                       output="data.frame")
   expect_equal(out$term[1:7],
-               c("(Intercept)", "", "mpg", "", "mpg × vs", "", "Num.Obs."))
+               c("(Intercept)", "(Intercept)", "mpg", "mpg", "mpg × vs", "mpg × vs", "Num.Obs."))
 })
 
 test_that("omit coefficients using regular expressions", {
@@ -18,7 +18,7 @@ test_that("omit coefficients using regular expressions", {
   mod$OLS <- lm(am ~ drat, data = mtcars)
   mod$Logit <- glm(am ~ qsec, data = mtcars, family = binomial())
 
-  raw <- modelsummary:::extract_models(mod, coef_omit = c('drat|qsec'))
+  raw <- modelsummary(mod, coef_omit = c('drat|qsec'), output="dataframe")
 
   truth <- c('(Intercept)', '(Intercept)', 'Num.Obs.')
   expect_equal(unname(raw[[2]][1:3]), truth)