+++ tests/testthat/test-coef_rename.R
@@ -9,7 +9,7 @@ mod$Logit <- glm(am ~ qsec, data = mtcars, family = binomial())
 test_that("rename 2 out of 3 coefficients", {
   cmap <- c('(Intercept)'='Constant', 'drat'='Rear axle ratio')
   mod <- msummary(mod, "dataframe", coef_rename=cmap)
-  known <- c("Constant", "", "Rear axle ratio", "", "qsec")
+  known <- c("Constant", "Constant", "Rear axle ratio", "Rear axle ratio", "qsec")
   expect_equal(mod$term[1:5], known)
 })