+++ tests/testthat/test-ellipses.R
@@ -5,9 +5,9 @@ library(modelsummary)
 test_that("exponentiate logit coefficients", {
 
   mod <- glm(am ~ mpg, mtcars, family = binomial)
-  raw <- modelsummary:::extract_models(mod, exponentiate = TRUE)
+  raw <- modelsummary(mod, output="dataframe", exponentiate=TRUE)
 
-  truth <- c("353.65253", "(76.04873)", "-57.54523", "(20.92205)", "", "")
-  expect_equal('1.359', unname(raw[[4]])[3])
+  truth <- c("0.001", "(2.351)", "1.359", "(0.115)", "32", "33.7", "36.6", "-14.838")
+  expect_equal(truth, unname(raw[[4]]))
 
 })