+++ tests/testthat/test-gt.R
@@ -11,6 +11,7 @@ models[['OLS 2']] <- lm(vs ~ hp + wt, mtcars)
 models[['Logit 1']] <- glm(vs ~ hp + drat, mtcars, family = binomial())
 models[['Logit 2']] <- glm(am ~ hp + disp, mtcars, family = binomial())
 
+
 test_that("gof_omit='.*' used to produce an error", {
 
   mod <- lm(mpg ~ wt, mtcars)
@@ -45,7 +46,7 @@ test_that("complex html table", {
       subtitle = 'Models estimated using the mtcars dataset.') %>%
     gt::as_raw_html()
 
-  expect_known_output(cat(raw), "known_output/complex_table.html")
+  expect_known_output(cat(raw), "known_output/complex_table.html", update=FALSE)
 
 })
 
@@ -53,7 +54,7 @@ test_that("title", {
 
   raw <- msummary(models, output = "gt", title = 'This is a title for my table.') %>%
     gt::as_raw_html()
-  expect_known_output(cat(raw), "known_output/title.html")
+  expect_known_output(cat(raw), "known_output/title.html", update=FALSE)
 
 })
 
@@ -70,6 +71,6 @@ test_that("background color", {
       locations = cells_body(columns = vars(`Logit 2`), rows = 2:6)) %>%
     as_raw_html()
 
-  expect_known_output(cat(raw), "known_output/background_color.html")
+  expect_known_output(cat(raw), "known_output/background_color.html", update=FALSE)
 
 })